Extract raftstore/store/observer subpackage; consolidate transitions
observer subpackage owns the post-apply event fanout. Same pattern as
the region extraction:
  observer.Runtime  - dispatcher; Register / Emit / Dropped / Close
  observer.Event    - the apply event payload (was ApplyEvent)
  observer.Source   - SourceCommit / SourceResolveLock
  observer.Observer - consumer interface
  observer.EventsFromCommand - extractor over (entry, req, resp)

Store keeps the existing public surface via type aliases:
  type ApplyEvent                = observer.Event
  type ApplyEventSource          = observer.Source
  type ApplyObserver             = observer.Observer
  type ApplyObserverRegistration = observer.Registration

External callers (raftstore/kv, fsmeta integration, raftstore
integration tests) keep importing storepkg.ApplyEvent unchanged.

Store-package consolidation alongside the extraction:
  - transition_builder.go + transition_executor.go merged into
    transitions.go. They were two halves of one subsystem (build →
    execute) with no independent reuse, and now form a single ~580 LoC
    file.
  - Test files merged into transitions_test.go.
  - Dead errRegionNotFound deleted (last call site moved into
    region.ErrNotFound when the region subpackage was extracted).
